requirements.txt
dotenv
redis
ollama
discord.py
PyNaCl